0.23077 correccion en bonos

+++ model/Bono.go
@@ -3,6 +3,6 @@ package model
 // Tbono es la estructura para usuario
 type Tbono struct {
 	ID       int64
-	Precio   int
+	Precio   float32
 	Sesiones int
 }
